+++ Classes/Configuration/ConfigurationFinder.php
@@ -11,12 +11,12 @@ declare(strict_types=1);
 
 namespace Brotkrueml\MatomoWidgets\Configuration;
 
-use Brotkrueml\MatomoWidgets\Adapter\YamlFileLoader;
 use Brotkrueml\MatomoWidgets\Domain\Entity\CustomDimension;
 use Brotkrueml\MatomoWidgets\Domain\Validation\CustomDimensionConfigurationValidator;
 use Brotkrueml\MatomoWidgets\Extension;
 use Symfony\Component\Finder\Exception\DirectoryNotFoundException;
 use Symfony\Component\Finder\Finder;
+use Symfony\Component\Yaml\Yaml;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
@@ -24,9 +24,15 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
  */
 final class ConfigurationFinder
 {
+    /**
+     * Regex pattern for allowed characters in environment variables
+     * @see https://regex101.com/r/hIXvef/1
+     * @see https://stackoverflow.com/questions/2821043/allowed-characters-in-linux-environment-variable-names
+     */
+    private const ENV_VAR_REGEX = '/^%env\(([a-zA-Z_]\w*)\)%$/';
+
     public static function buildConfigurations(string $configPath, bool $isMatomoIntegrationAvailable): Configurations
     {
-        $yamlFileLoader = YamlFileLoader::get();
         $configurationsArray = [];
         foreach (self::getConfigurationFiles($configPath) as $file) {
             $filePath = $file->getPathname();
@@ -34,22 +40,24 @@ final class ConfigurationFinder
                 continue;
             }
             $filePath = GeneralUtility::fixWindowsFilePath($filePath);
-            $siteConfiguration = $yamlFileLoader->load($filePath);
+            $siteConfiguration = Yaml::parseFile($filePath);
 
             $considerMatomoIntegration = $isMatomoIntegrationAvailable
                 && ($siteConfiguration['matomoWidgetsConsiderMatomoIntegration'] ?? false);
 
             if ($considerMatomoIntegration) {
                 $url = $siteConfiguration['matomoIntegrationUrl'] ?? '';
-                $idSite = (int) ($siteConfiguration['matomoIntegrationSiteId'] ?? 0);
+                $idSite = (string) ($siteConfiguration['matomoIntegrationSiteId'] ?? 0);
             } else {
                 $url = $siteConfiguration['matomoWidgetsUrl'] ?? '';
-                $idSite = (int) ($siteConfiguration['matomoWidgetsIdSite'] ?? 0);
+                $idSite = (string) ($siteConfiguration['matomoWidgetsIdSite'] ?? 0);
             }
-            if (\str_starts_with((string) $url, '//')) {
+            $url = self::resolveEnvironmentVariable($url);
+            if (\str_starts_with($url, '//')) {
                 // We have a relative URL: prepend the scheme from the base URL of the site configuration
                 $url = \parse_url((string) $siteConfiguration['base'], \PHP_URL_SCHEME) . ':' . $url;
             }
+            $idSite = (int) self::resolveEnvironmentVariable($idSite);
             if ($url === '') {
                 continue;
             }
@@ -62,9 +70,10 @@ final class ConfigurationFinder
             } else {
                 $pagesNotFoundTemplate = $siteConfiguration['matomoWidgetsPagesNotFoundTemplate'] ?? '';
             }
+            $pagesNotFoundTemplate = self::resolveEnvironmentVariable($pagesNotFoundTemplate);
 
-            $siteTitle = $siteConfiguration['matomoWidgetsTitle'] ?? '';
-            $tokenAuth = $siteConfiguration['matomoWidgetsTokenAuth'] ?? '';
+            $siteTitle = self::resolveEnvironmentVariable($siteConfiguration['matomoWidgetsTitle'] ?? '');
+            $tokenAuth = self::resolveEnvironmentVariable($siteConfiguration['matomoWidgetsTokenAuth'] ?? '');
 
             $pathSegments = \explode('/', $file->getPath());
             $identifier = \end($pathSegments);
@@ -76,7 +85,7 @@ final class ConfigurationFinder
 
             $activeWidgets = GeneralUtility::trimExplode(
                 ',',
-                $siteConfiguration['matomoWidgetsActiveWidgets'] ?? '',
+                self::resolveEnvironmentVariable($siteConfiguration['matomoWidgetsActiveWidgets'] ?? ''),
                 true,
             );
             $customDimensions = self::buildCustomDimensions($siteConfiguration['matomoWidgetsCustomDimensions'] ?? []);
@@ -122,6 +131,24 @@ final class ConfigurationFinder
         }
 
         return [...$siteFiles, ...$additionalFiles];
+    }
+
+    /**
+     * This method is necessary as environment variables are not resolved when configuration
+     * is available in controllers.
+     */
+    private static function resolveEnvironmentVariable(string $value): string
+    {
+        if (\preg_match(self::ENV_VAR_REGEX, $value, $matches) !== 1) {
+            return $value;
+        }
+
+        $resolvedValue = \getenv($matches[1]);
+        if ($resolvedValue === false) {
+            return '';
+        }
+
+        return $resolvedValue;
     }
 
     /**